Add QuadA, QuadB, QuadC, QuadD, and QuadE functions
This commit adds five new functions to the main package: QuadA, QuadB, QuadC, QuadD, and QuadE. QuadA is a function that generates a matrix with a specific pattern and prints it. QuadB, QuadC, QuadD, and QuadE are empty functions that currently do nothing. These functions will be used for future development.